test: FakeActivity들을 테스트함수 내부에서 선언하도록 바꾸고 Fake뷰모델들의 네이밍에 테스트 의도를 담도록 수정
