+++ alsonglibrary2/src/test/java/com/example/alsonglibrary2/AutoDIManagerTest.kt
@@ -1,17 +1,19 @@
 package com.example.alsonglibrary2
 
+import androidx.appcompat.app.AppCompatActivity
 import com.example.alsonglibrary2.di.AutoDIManager
+import com.example.alsonglibrary2.di.createAutoDIViewModel
 import com.example.alsonglibrary2.fixtures.QualifiedDependencyProvider
-import com.example.alsonglibrary2.fixtures.activity.FakeActivity1
-import com.example.alsonglibrary2.fixtures.activity.FakeActivity2
-import com.example.alsonglibrary2.fixtures.activity.FakeActivity3
-import com.example.alsonglibrary2.fixtures.activity.FakeActivity4
-import com.example.alsonglibrary2.fixtures.activity.FakeActivity5
-import com.example.alsonglibrary2.fixtures.activity.FakeActivity6
-import com.example.alsonglibrary2.fixtures.activity.FakeActivity7
 import com.example.alsonglibrary2.fixtures.instance.defaultFakeRepository1
 import com.example.alsonglibrary2.fixtures.instance.defaultFakeRepository2
 import com.example.alsonglibrary2.fixtures.repository.FakeRepository
+import com.example.alsonglibrary2.fixtures.viewmodel.ConstructorInjectedWithQualifierFakeViewModel
+import com.example.alsonglibrary2.fixtures.viewmodel.ConstructorInjectedWithTwoQualifierFakeViewModel
+import com.example.alsonglibrary2.fixtures.viewmodel.ConstructorInjectedWithoutQualifierFakeViewModel
+import com.example.alsonglibrary2.fixtures.viewmodel.FieldInjectedWithQualifierAndNoProviderFakeViewModel
+import com.example.alsonglibrary2.fixtures.viewmodel.FieldInjectedWithQualifierFakeViewModel
+import com.example.alsonglibrary2.fixtures.viewmodel.FieldInjectedWithTwoQualifierFakeViewModel
+import com.example.alsonglibrary2.fixtures.viewmodel.FieldInjectedWithoutQualifierFakeViewModel
 import com.google.common.truth.Truth.assertThat
 import org.junit.Before
 import org.junit.Test
@@ -30,6 +32,10 @@ class AutoDIManagerTest {
     fun `생성자에 등록된 의존성을 주입할 수 있다`() {
         // given
         AutoDIManager.registerDependency<FakeRepository>(defaultFakeRepository1)
+
+        class FakeActivity1 : AppCompatActivity() {
+            val viewModel by createAutoDIViewModel<ConstructorInjectedWithoutQualifierFakeViewModel>()
+        }
         val activity = Robolectric.buildActivity(FakeActivity1::class.java).create().get()
 
         // then
@@ -39,7 +45,10 @@ class AutoDIManagerTest {
 
     @Test(expected = Exception::class)
     fun `등록된 의존성이 존재하지 않으면 의존성 주입에 실패한다`() {
-        // given - Qualifier가 없는 생성자 주입
+        // given
+        class FakeActivity1 : AppCompatActivity() {
+            val viewModel by createAutoDIViewModel<ConstructorInjectedWithoutQualifierFakeViewModel>()
+        }
         val activity = Robolectric.buildActivity(FakeActivity1::class.java).create().get()
 
         // then
@@ -50,6 +59,10 @@ class AutoDIManagerTest {
     fun `생성자에 Qualifier로 지정된 의존성을 주입할 수 있다`() {
         // given
         AutoDIManager.provider = QualifiedDependencyProvider
+
+        class FakeActivity2 : AppCompatActivity() {
+            val viewModel by createAutoDIViewModel<ConstructorInjectedWithQualifierFakeViewModel>()
+        }
         val activity = Robolectric.buildActivity(FakeActivity2::class.java).create().get()
 
         // then
@@ -61,8 +74,11 @@ class AutoDIManagerTest {
     fun `필드에 등록된 의존성을 주입할 수 있다`() {
         // given
         AutoDIManager.registerDependency<FakeRepository>(defaultFakeRepository1)
-        val activity = Robolectric.buildActivity(FakeActivity3::class.java).create().get()
 
+        class FakeActivity3 : AppCompatActivity() {
+            val viewModel by createAutoDIViewModel<FieldInjectedWithoutQualifierFakeViewModel>()
+        }
+        val activity = Robolectric.buildActivity(FakeActivity3::class.java).create().get()
 
         // then
         assertThat(activity.viewModel).isNotNull()
@@ -73,6 +89,10 @@ class AutoDIManagerTest {
     fun `필드에 Qualifier로 지정된 의존성을 주입할 수 있다`() {
         // given
         AutoDIManager.provider = QualifiedDependencyProvider
+
+        class FakeActivity4 : AppCompatActivity() {
+            val viewModel by createAutoDIViewModel<FieldInjectedWithQualifierFakeViewModel>()
+        }
         val activity = Robolectric.buildActivity(FakeActivity4::class.java).create().get()
 
         // then
@@ -84,6 +104,10 @@ class AutoDIManagerTest {
     fun `의존성이 등록되어있지 않고 Qualifier로 지정된 의존성도 없으면 의존성 주입이 실패한다`() {
         // given
         AutoDIManager.provider = QualifiedDependencyProvider
+
+        class FakeActivity5 : AppCompatActivity() {
+            val viewModel by createAutoDIViewModel<FieldInjectedWithQualifierAndNoProviderFakeViewModel>()
+        }
         val activity = Robolectric.buildActivity(FakeActivity5::class.java).create().get()
 
         // then
@@ -94,6 +118,10 @@ class AutoDIManagerTest {
     fun `필드에 같은 타입의 의존성이 여러개 있을 때 Qualifier로 인스턴스를 다르게 주입할 수 있다`() {
         // given
         AutoDIManager.provider = QualifiedDependencyProvider
+
+        class FakeActivity6 : AppCompatActivity() {
+            val viewModel by createAutoDIViewModel<FieldInjectedWithTwoQualifierFakeViewModel>()
+        }
         val activity = Robolectric.buildActivity(FakeActivity6::class.java).create().get()
 
         // then
@@ -106,6 +134,10 @@ class AutoDIManagerTest {
     fun `생성자에 같은 타입의 의존성이 여러개 있을 때 Qualifier로 인스턴스를 다르게 주입할 수 있다`() {
         // given
         AutoDIManager.provider = QualifiedDependencyProvider
+
+        class FakeActivity7 : AppCompatActivity() {
+            val viewModel by createAutoDIViewModel<ConstructorInjectedWithTwoQualifierFakeViewModel>()
+        }
         val activity = Robolectric.buildActivity(FakeActivity7::class.java).create().get()
 
         // then
