+++ alsonglibrary2/src/test/java/com/example/alsonglibrary2/fixtures/viewmodel/ConstructorInjectedWithTwoQualifierFakeViewModel.kt
@@ -1,17 +1,11 @@
 package com.example.alsonglibrary2.fixtures.viewmodel
 
 import androidx.lifecycle.ViewModel
-import com.example.alsonglibrary2.di.FieldInject
 import com.example.alsonglibrary2.fixtures.DefaultRepository1
 import com.example.alsonglibrary2.fixtures.DefaultRepository2
 import com.example.alsonglibrary2.fixtures.repository.FakeRepository
 
-class FakeViewModel6 : ViewModel() {
-    @FieldInject
-    @DefaultRepository1
-    lateinit var fakeRepository1: FakeRepository
-
-    @FieldInject
-    @DefaultRepository2
-    lateinit var fakeRepository2: FakeRepository
-}
+class ConstructorInjectedWithTwoQualifierFakeViewModel(
+    @DefaultRepository1 val fakeRepository1: FakeRepository,
+    @DefaultRepository2 val fakeRepository2: FakeRepository,
+) : ViewModel()
