+++ alsonglibrary2/src/test/java/com/example/alsonglibrary2/fixtures/viewmodel/ConstructorInjectedWithoutQualifierFakeViewModel.kt
@@ -3,6 +3,6 @@ package com.example.alsonglibrary2.fixtures.viewmodel
 import androidx.lifecycle.ViewModel
 import com.example.alsonglibrary2.fixtures.repository.FakeRepository
 
-class FakeViewModel1(
+class ConstructorInjectedWithoutQualifierFakeViewModel(
     val fakeRepository: FakeRepository,
 ) : ViewModel()
