+++ alsonglibrary2/src/test/java/com/example/alsonglibrary2/fixtures/viewmodel/FieldInjectedWithQualifierAndNoProviderFakeViewModel.kt
@@ -2,11 +2,11 @@ package com.example.alsonglibrary2.fixtures.viewmodel
 
 import androidx.lifecycle.ViewModel
 import com.example.alsonglibrary2.di.FieldInject
-import com.example.alsonglibrary2.fixtures.DefaultRepository1
+import com.example.alsonglibrary2.fixtures.DefaultRepository0
 import com.example.alsonglibrary2.fixtures.repository.FakeRepository
 
-class FakeViewModel4 : ViewModel() {
+class FieldInjectedWithQualifierAndNoProviderFakeViewModel : ViewModel() {
     @FieldInject
-    @DefaultRepository1
+    @DefaultRepository0
     lateinit var fakeRepository: FakeRepository
 }
