+++ alsonglibrary2/src/test/java/com/example/alsonglibrary2/fixtures/viewmodel/FieldInjectedWithQualifierFakeViewModel.kt
@@ -1,9 +1,12 @@
 package com.example.alsonglibrary2.fixtures.viewmodel
 
 import androidx.lifecycle.ViewModel
+import com.example.alsonglibrary2.di.FieldInject
 import com.example.alsonglibrary2.fixtures.DefaultRepository1
 import com.example.alsonglibrary2.fixtures.repository.FakeRepository
 
-class FakeViewModel2(
-    @DefaultRepository1 val fakeRepository: FakeRepository,
-) : ViewModel()
+class FieldInjectedWithQualifierFakeViewModel : ViewModel() {
+    @FieldInject
+    @DefaultRepository1
+    lateinit var fakeRepository: FakeRepository
+}
