+++ alsonglibrary2/src/test/java/com/example/alsonglibrary2/fixtures/viewmodel/FieldInjectedWithoutQualifierFakeViewModel.kt
@@ -4,7 +4,7 @@ import androidx.lifecycle.ViewModel
 import com.example.alsonglibrary2.di.FieldInject
 import com.example.alsonglibrary2.fixtures.repository.FakeRepository
 
-class FakeViewModel3 : ViewModel() {
+class FieldInjectedWithoutQualifierFakeViewModel : ViewModel() {
     @FieldInject
     lateinit var fakeRepository: FakeRepository
 }
